# lib/api-zod/src/generated/types/campaignCost.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GlamNet API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignCost {
  artists: number;
  artistFees: number;
  fee: number;
  total: number;
  deposit: number;
  balance: number;
}