# lib/api-zod/src/generated/types/campaignPaymentRecord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GlamNet API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignPaymentRecordKind } from './campaignPaymentRecordKind';
import type { CampaignPaymentRecordStatus } from './campaignPaymentRecordStatus';

export interface CampaignPaymentRecord {
  id: string;
  kind: CampaignPaymentRecordKind;
  amount: number;
  status: CampaignPaymentRecordStatus;
  createdAt: Date;
  paidAt: Date | null;
}