# lib/api-zod/src/generated/types/campaignSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GlamNet API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignCost } from './campaignCost';
import type { CampaignJob } from './campaignJob';
import type { CampaignPaymentRecord } from './campaignPaymentRecord';
import type { CampaignSummaryStage } from './campaignSummaryStage';
import type { CampaignTeamMember } from './campaignTeamMember';
import type { CastingCall } from './castingCall';

export interface CampaignSummary {
  call: CastingCall;
  stage: CampaignSummaryStage;
  team: CampaignTeamMember[];
  jobs: CampaignJob[];
  payments: CampaignPaymentRecord[];
  cost: CampaignCost;
  paid: number;
  outstanding: number;
  balanceDueDate: string | null;
  balanceOverdue: boolean;
  depositAllowed: boolean;
  canFund: boolean;
  canPayBalance: boolean;
  blockers: string[];
}