# lib/api-zod/src/generated/types/castingCall.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GlamNet API
 * OpenAPI spec version: 0.1.0
 */
import type { CastingCallMyStatus } from './castingCallMyStatus';
import type { CastingCallStatus } from './castingCallStatus';

export interface CastingCall {
  id: string;
  brandId: string;
  brandName: string;
  title: string;
  brief: string;
  budget: string;
  deadline: string;
  specialty: string;
  applicantCount: number;
  hasApplied?: boolean;
  createdAt: string;
  artistsNeeded: number;
  /** What each artist is paid, in rand. She keeps all of it. */
  ratePerArtist: number;
  eventDate: string | null;
  eventTime: string;
  location: string;
  status: CastingCallStatus;
  spotsFilled: number;
  brandVerified: boolean;
  /** Where the signed-in artist stands on this campaign. */
  myStatus?: CastingCallMyStatus;
}