# lib/api-zod/src/generated/types/ownerPendingBrand.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GlamNet API
 * OpenAPI spec version: 0.1.0
 */

export interface OwnerPendingBrand {
  profileId: string;
  userId: string;
  companyName: string;
  registrationNumber: string | null;
  vatNumber: string | null;
  website: string | null;
  billingAddress: string | null;
  contactName: string;
  email: string;
  phone: string | null;
  submittedAt: Date | null;
}