# lib/api-zod/src/generated/types/startCampaignPaymentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GlamNet API
 * OpenAPI spec version: 0.1.0
 */
import type { StartCampaignPaymentInputKind } from './startCampaignPaymentInputKind';

export interface StartCampaignPaymentInput {
  kind: StartCampaignPaymentInputKind;
}